refactor: introduce shared root layout
