+++ src/app/(home)/home/page.tsx
@@ -1,4 +1,9 @@
 import React from 'react';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'Home',
+};
 
 export default function Home() {
   return (
